# Fixed Unit Test salt-max in case of optimized kernel, with hash-type 22 and 23