--- a/tools/test_modules/m00023.pm
+++ b/tools/test_modules/m00023.pm
@@ -10,7 +10,7 @@ use warnings;
 
 use Digest::MD5 qw (md5_hex);
 
-sub module_constraints { [[0, 247], [0, 247], [0, 47], [0, 47], [0, 47]] }
+sub module_constraints { [[0, 247], [0, 247], [0, 47], [0, 43], [0, 47]] }
 
 sub module_generate_hash
 {